Move kmer.rs from bitpacked_kmer_sorting to crate.
